reorg jobs
