+++ internal/domain/track.go
@@ -1,5 +1,6 @@
 package domain
 
+// Track represents an individual track in a tracklist.
 type Track struct {
 	Name        string `json:"name"`
 	Artist      string `json:"artist"`
@@ -8,10 +9,11 @@ type Track struct {
 	TrackNumber int    `json:"track_number"`
 }
 
+// Tracklist represents a collection of tracks in a DJ set.
 type Tracklist struct {
 	Name   string   `json:"name"`
-	Year   int      `json:"year"`
+	Year   int      `json:"year,omitempty"`
 	Artist string   `json:"artist"`
-	Genre  string   `json:"genre"`
+	Genre  string   `json:"genre,omitempty"`
 	Tracks []*Track `json:"tracks"`
 }
